Matrix-of-Truth-main/Matrix-of-Truth-main/frontend_extension_matrix/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0
google-generativeai>=0.3.0
gunicorn>=21.2.0
httpx>=0.27.0

